+++ forgerock-authenticator/CHANGELOG.md
@@ -1,4 +1,13 @@
-# Version 1.1.2
+# Version 1.2.0
+
+##[1.2.0]
+#### Added
+- New Authenticator Policies feature [SDKS-2326]
+
+#### Updated
+
+- Android and iOS `forgerock-authenticator` dependency to the 4.0.0-beta4 release.
+- Sample app updated to use the new Authenticator policies feature
 
 ## [1.1.2]